Rename hook to `useIItemNo.tsx`
