+++ packages/dnb-eufemia/src/extensions/forms/FieldBlock/FieldBlock.tsx
@@ -42,7 +42,7 @@ import SubmitIndicator from '../Form/SubmitIndicator/SubmitIndicator'
 import { createSharedState } from '../../../shared/helpers/useSharedState'
 import useTranslation from '../hooks/useTranslation'
 import { FormError } from '../utils'
-import { useIterateItemNo } from '../Iterate/ItemNo/useIterateNo'
+import { useIterateItemNo } from '../Iterate/ItemNo/useIItemNo'
 
 export const states: Array<StateTypes> = ['error', 'info', 'warning']
