+++ packages/dnb-eufemia/src/extensions/forms/Iterate/ItemNo/__tests__/useIterateItemNo.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import { useIterateItemNo } from '../useIterateNo'
+import { useIterateItemNo } from '../useIItemNo'
 import IterateItemContext from '../../IterateItemContext'
 
 function TestComponent(props: any) {
